implements static file serving
Serverside rendering comes in handy especially when its about serving
static assets. Since we always have an express server running in the
background its very simple to have a clean way of serving static assets.

The user now has the possibility to create a /static/ folder inside the
root folder of the project. all files that are inside the folder will be
copied to the build folder with CopyWebpackPlugin and the served via
express.static. This is solved through a conditional plugin which only
get's used of the user actually has specified a /static/ folder.