restructure readme
the readme got a new structure. it now starts with a table of contents
that brings you to the main sections of the readme. i've cleanup the
'Getting started' section and added a section on serving static files.
the overall idea was moved down to a less prominted place, to ensure
that people can get started right away without being distracted by
understanding the underlying idea of the rendering setup.

- fix #8